feat: Add NotifyRewards component for displaying reward notifications
