+++ frontend/components/Main/MainRewards.tsx
@@ -1,18 +1,12 @@
-import { Button, Col, Flex, Modal, Row, Skeleton, Tag, Typography } from 'antd';
-import Image from 'next/image';
-import { useCallback, useEffect, useState } from 'react';
+import { Col, Flex, Row, Skeleton, Tag, Typography } from 'antd';
 import styled from 'styled-components';
 
 import { balanceFormat } from '@/common-util';
 import { COLOR } from '@/constants';
 import { useBalance } from '@/hooks';
-import { useElectronApi } from '@/hooks/useElectronApi';
 import { useReward } from '@/hooks/useReward';
-import { useStore } from '@/hooks/useStore';
 
-import { ConfettiAnimation } from '../common/ConfettiAnimation';
-
-const { Text, Title } = Typography;
+const { Text } = Typography;
 
 const RewardsRow = styled(Row)`
   margin: 0 -24px;
@@ -31,7 +25,7 @@ const Loader = () => (
   </Flex>
 );
 
-const DisplayRewards = () => {
+export const MainRewards = () => {
   const {
     availableRewardsForEpochEth,
     isEligibleForRewards,
@@ -87,126 +81,3 @@ const DisplayRewards = () => {
     </RewardsRow>
   );
 };
-
-const SHARE_TEXT = `I just earned my first reward through the Operate app powered by #olas!\n\nDownload the Pearl app:`;
-const OPERATE_URL = 'https://olas.network/operate?pearl=first-reward';
-
-const NotifyRewards = () => {
-  const { isEligibleForRewards, availableRewardsForEpochEth } = useReward();
-  const { totalOlasBalance } = useBalance();
-  const { showNotification, store } = useElectronApi();
-  const { storeState } = useStore();
-
-  const [canShowNotification, setCanShowNotification] = useState(false);
-
-  // hook to set the flag to show the notification
-  useEffect(() => {
-    if (!isEligibleForRewards) return;
-    if (!storeState) return;
-    if (storeState?.firstRewardNotificationShown) return;
-    if (!availableRewardsForEpochEth) return;
-
-    setCanShowNotification(true);
-  }, [
-    isEligibleForRewards,
-    availableRewardsForEpochEth,
-    showNotification,
-    storeState,
-  ]);
-
-  // hook to show desktop app notification
-  useEffect(() => {
-    if (!canShowNotification) return;
-
-    showNotification?.(
-      'Your agent earned its first staking rewards!',
-      `Congratulations! Your agent just got the first reward for you! Your current balance: ${availableRewardsForEpochEth} OLAS`,
-    );
-  }, [canShowNotification, availableRewardsForEpochEth, showNotification]);
-
-  const closeNotificationModal = useCallback(() => {
-    setCanShowNotification(false);
-
-    // once the notification is closed, set the flag to true
-    store?.set?.('firstRewardNotificationShown', true);
-  }, [store]);
-
-  const onTwitterShare = useCallback(() => {
-    const encodedText = encodeURIComponent(SHARE_TEXT);
-    const encodedURL = encodeURIComponent(OPERATE_URL);
-
-    window.open(
-      `https://twitter.com/intent/tweet?text=${encodedText}&url=${encodedURL}`,
-      '_blank',
-    );
-  }, []);
-
-  if (!canShowNotification) return null;
-
-  return (
-    <Modal
-      open={canShowNotification}
-      width={400}
-      onCancel={closeNotificationModal}
-      footer={[
-        <Button
-          key="back"
-          type="primary"
-          block
-          size="large"
-          className="mt-8"
-          onClick={onTwitterShare}
-        >
-          <Flex align="center" justify="center" gap={2}>
-            Share on
-            <Image
-              src="/twitter.svg"
-              width={24}
-              height={24}
-              alt="Share on twitter"
-            />
-          </Flex>
-        </Button>,
-      ]}
-    >
-      <ConfettiAnimation />
-
-      <Flex align="center" justify="center">
-        <Image
-          src="/splash-robot-head.png"
-          width={100}
-          height={100}
-          alt="OLAS logo"
-        />
-      </Flex>
-
-      <Title level={5} className="mt-12">
-        Your agent just earned the first reward!
-      </Title>
-
-      <Flex vertical gap={16}>
-        <Text>
-          Congratulations! Your agent just earned the first
-          <Text strong>
-            {` ${balanceFormat(availableRewardsForEpochEth, 2)} OLAS `}
-          </Text>
-          for you!
-        </Text>
-
-        <Text>
-          Your current balance:
-          <Text strong>{` ${balanceFormat(totalOlasBalance, 2)} OLAS `}</Text>
-        </Text>
-
-        <Text>Keep it running to get even more!</Text>
-      </Flex>
-    </Modal>
-  );
-};
-
-export const MainRewards = () => (
-  <>
-    <DisplayRewards />
-    <NotifyRewards />
-  </>
-);
